+++ sdmx-dl-api/src/test/java/tests/sdmxdl/api/RepoSamples.java
@@ -49,9 +49,9 @@ public class RepoSamples {
     public static final Codelist CL3 = Codelist.builder().ref(CL_REF3).code("INDUSTRY", "Industry").code("XXX", "Other").build();
     public static final Codelist CL4 = Codelist.builder().ref(CL_REF4).code("A", "Normal value").build();
 
-    public static final Dimension DIM1 = Dimension.builder().id("FREQ").codelist(CL1).name("Frequency").position(1).build();
-    public static final Dimension DIM2 = Dimension.builder().id("REGION").codelist(CL2).name("Region").position(3).build();
-    public static final Dimension DIM3 = Dimension.builder().id("SECTOR").codelist(CL3).name("Sector").position(4).build();
+    public static final Dimension DIM1 = Dimension.builder().id("FREQ").codelist(CL1).name("Frequency").build();
+    public static final Dimension DIM2 = Dimension.builder().id("REGION").codelist(CL2).name("Region").build();
+    public static final Dimension DIM3 = Dimension.builder().id("SECTOR").codelist(CL3).name("Sector").build();
 
     public static final Attribute NOT_CODED_ATTRIBUTE = Attribute.builder().id("TITLE").name("Title").relationship(AttributeRelationship.SERIES).build();
     public static final Attribute CODED_ATTRIBUTE = Attribute.builder().id("OBS_STATUS").codelist(CL4).name("Observation status").relationship(AttributeRelationship.OBSERVATION).build();