+++ CHANGELOG.md
@@ -28,6 +28,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - ![API] Modify TimeInterval to closely follow ISO-8601
 - ![API] Simplify listeners
 - ![API] Retrieve testing connection URI for external usage [#1058](https://github.com/nbbrd/sdmx-dl/issues/1058)
+- ![API] Remove confusing Dimension#getPosition()
 - ![SOURCE] Update ABS endpoint [#1037](https://github.com/nbbrd/sdmx-dl/issues/1037)
 - ![SOURCE] Remove TNSO source [#1036](https://github.com/nbbrd/sdmx-dl/issues/1036)
 - ![SOURCE] Remove SE source [#1057](https://github.com/nbbrd/sdmx-dl/issues/1057)